fix slice data type

+++ mxgan/viz.py
@@ -8,8 +8,8 @@ def _fill_buf(buf, i, img, shape):
     n = buf.shape[0]/shape[1]
     m = buf.shape[1]/shape[0]
 
-    sx = (i%m)*shape[0]
-    sy = (i//m)*shape[1]
+    sx = int((i%m)*shape[0])
+    sy = int((i//m)*shape[1])
     buf[sy:sy+shape[1], sx:sx+shape[0], :] = img
